+++ ui/src/App.jsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 
 // import Htop from './components/Htop';
-import FileUpload from './components/FileUpload'
+import Htop from './components/Htop'
 import './App.css';
 
 
@@ -10,7 +10,7 @@ class App extends Component {
     return (
       <div>
         <div className="App">
-          <FileUpload/>
+          <Htop/>
         </div>
       </div>
     );
